Rename StyledBungieImage to BungieImageAndAmmo
