+++ src/app/d2-loadout-builder/locked-armor/popup/SelectableBungieImage.tsx
@@ -3,7 +3,7 @@ import * as classNames from 'classnames';
 import { t } from 'i18next';
 import * as React from 'react';
 import { LockedItemType } from '../../types';
-import StyledBungieImage from '../../../dim-ui/StyledBungieImage';
+import StyledBungieImage from '../../../dim-ui/BungieImageAndAmmo';
 
 export default function SelectableBungieImage({
   perk,
