+++ src/app/item-popup/ItemSockets.tsx
@@ -14,7 +14,7 @@ import { curationsSelector, getInventoryCuratedRoll } from '../curated-rolls/red
 import { RootState } from '../store/reducers';
 import { dimDestinyTrackerService } from '../item-review/destiny-tracker.service';
 import { $rootScope } from 'ngimport';
-import StyledBungieImage from '../dim-ui/StyledBungieImage';
+import StyledBungieImage from '../dim-ui/BungieImageAndAmmo';
 
 interface ProvidedProps {
   item: D2Item;
